Remove previously laid foundations
These foundations may not be suitable for the newly assigned thesis
homework. Plus the board was implemented with arrays, which is clearly
an unfortunate thing when you remember that you are writing the program
in C++, not C.

So I'm going to start all over again. I don't think I'm collaborating
with my classmate anymore, so I modified the README file as well.

Let's start all over again!