Working on storage
1. Added some functions to the Board type. I think the stage is basically finished now, so what I have to do next is to think about the interface.

2. Added the bones for an Interface header.

I think I will use a base class anyways. I don't know where I'll put the algorithms for Renju really, but I assume I will put it in the interface. I am not sure if I will use inheritance or just include it as an element in the Interface.
The interface is called an interface and singled out for it to have the following abilities:

1.It can Output the board to either the terminal or the GUI (if I get time to implement one)

2. It can get input;

3. It can parse the input and turn it into something I can see on the board;

4. It can decide whether the input was valid, and if the move was winning.

I'm thinking the latter two may be better implemented on the base class. The algorithm will be the hardest thing.

I also need to consider how to do the multi-threading. I might actually need something to listen on the input channel, I'm not very sure. Though I thought that might actually be the job that a simple function could do? I don't know.